changed distance type from double to int, CentrosAdapter and partial implementation of the consultarDatos method in MainActivity

+++ app/src/main/java/com/example/proyectoa_pmdm_t2_junzhou/FilterDialog.java
@@ -46,7 +46,7 @@ public class FilterDialog extends DialogFragment {
                     Toast.makeText(getActivity(), "No puede haber campos vacíos", Toast.LENGTH_SHORT).show();
                 } else {
                     if(listener != null) {
-                        listener.onDatosListener(Double.parseDouble(lat), Double.parseDouble(lon), Double.parseDouble(dist));
+                        listener.onDatosListener(Double.parseDouble(lat), Double.parseDouble(lon), Integer.parseInt(dist));
                     }
                 }
